requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-magic==0.4.26
pytz==2022.1
sqlparse==0.4.2
